+++ demo/storybook/media/icon_sprite.story.exs
@@ -1,0 +1,4 @@
+defmodule Storybook.Components.IconSprite do
+  use PhoenixStorybook.Story, :component
+  use Doggo.Storybook, module: DemoWeb.CoreComponents, name: :icon_sprite
+end