+++ lib/doggo/components/steps.ex
@@ -13,6 +13,17 @@ defmodule Doggo.Components.Steps do
   end
 
   @impl true
+  def builder_doc do
+    """
+    - `:current_class` - This class is added to the current step.
+    - `:completed_class` - This class is added to previous steps.
+    - `:upcoming_class` - This class is added to upcoming steps.
+    - `:visually_hidden_class` - This class is used to visually hide the
+      accessibility text added to completed steps.
+    """
+  end
+
+  @impl true
   def usage do
     """
     With patch navigation: