+++ lib/doggo/storybook/date.ex
@@ -21,11 +21,30 @@ defmodule Doggo.Storybook.Date do
       },
       %Variation{
         id: :title_formatter,
-        description: "Hover over the date to see the title.",
+        description: "Hover over the date to see the title",
         attributes: %{
           title_formatter: &Date.to_gregorian_days/1,
           value: ~D[2023-02-05]
         }
+      },
+      %Variation{
+        id: :timezone,
+        description: "Shift DateTime to a different time zone",
+        attributes: %{
+          value: ~U[2023-02-05 23:22:05Z],
+          timezone:
+            if time_zone_db_configured?() do
+              "Asia/Tokyo"
+            end
+        },
+        template:
+          if !time_zone_db_configured?() do
+            """
+            <p>
+              This example requires a <a href="https://hexdocs.pm/elixir/DateTime.html#module-time-zone-database">time zone database to be configured</a>.
+            </p>
+            """
+          end
       }
     ]
   end
@@ -34,5 +53,10 @@ defmodule Doggo.Storybook.Date do
     %{
       attributes: %{value: ~D[2023-02-05]}
     }
+  end
+
+  defp time_zone_db_configured? do
+    Application.get_env(:elixir, :time_zone_database) !=
+      Calendar.UTCOnlyTimeZoneDatabase
   end
 end