+++ CHANGELOG.md
@@ -47,6 +47,11 @@
 - Set `sprite_url` as a compile time option.
 - Mark both components as `refining`.
 
+#### Date component
+
+- Improve story and documentation.
+- Mark component as `refining`.
+
 ## [0.7.0] - 2024-07-24
 
 ### Changed